refactor(song-management): #54  add trait  identify
